+++ maps/tiled.js
@@ -9,8 +9,15 @@ var olmMapFormat = {
     name: "Of Lands Map format",
     extension: "olm",
 
+    outputFiles: function(map, fileName) {
+        const baseName = fileName.substring(0, fileName.lastIndexOf("."));
+        return [baseName + "olm", baseName + "oldata"];
+    },
+
     write: function(map, fileName) {
         // base data
+        const baseName        = fileName.substring(0, fileName.lastIndexOf("."));
+        const baseDir         = fileName.substring(0, fileName.lastIndexOf("/"));
         var tilesets          = map.usedTilesets();
         var tilesets_dict     = tilesets.reduce(function(dict, x) {
                                     dict[x.name] = x;
@@ -30,9 +37,9 @@ var olmMapFormat = {
 
         var layer = map.layerAt(0); // terrain
         if (layer.isTileLayer) {
-            for (y = 0; y < layer.height; ++y) {
+            for (var y = 0; y < layer.height; ++y) {
                 out = out + "    [";
-                for (x = 0; x < layer.width; ++x)
+                for (var x = 0; x < layer.width; ++x)
                     out = out + layer.cellAt(x, y).tileId + ",";
                 out = out + "],\n";
             }
@@ -43,9 +50,9 @@ var olmMapFormat = {
 
         var layer = map.layerAt(1); // locations
         if (layer.isTileLayer) {
-            for (y = 0; y < layer.height; ++y) {
+            for (var y = 0; y < layer.height; ++y) {
                 out = out + "    [";
-                for (x = 0; x < layer.width; ++x)
+                for (var x = 0; x < layer.width; ++x)
                     out = out + layer.cellAt(x, y).tileId + ",";
                 out = out + "],\n";
             }
@@ -54,9 +61,61 @@ var olmMapFormat = {
         out = out + "]";
 
         // the main .olm file
-        var file = new TextFile(fileName, TextFile.WriteOnly);
-        file.write(out);
-        file.commit();
+        var olmFile = new TextFile(baseName + ".olm", TextFile.WriteOnly);
+        olmFile.write(out);
+        olmFile.commit();
+
+        // location .oldata parsing
+        var locdata = "";
+        for (var i = 0; i < tilesets_dict["Locations"].tiles.length; ++i) {
+            var tile      = tilesets_dict["Locations"].tiles[i];
+            var tile_data = tile.properties();
+
+            if (Object.keys(tile_data).length > 0) { // skips locations without properties
+                locdata += "[tile." + tile.id + "]\n";
+
+                for (var key in tile_data) {
+                    var val = typeof tile_data[key] === "string"  ? "\"" + tile_data[key] + "\""
+                            : typeof tile_data[key] === "number"  ? tile_data[key]
+                            : typeof tile_data[key] === "boolean" ? tile_data[key]
+                            : "";
+                    locdata += key + " = " + val + "\n";
+                }
+            }
+        }
+
+        // .oldata for locations
+        if (locdata.length > 0) { // skips if no properties are set (= user wants to make .oldata manually)
+            var locdataFile = new TextFile(baseDir + "/../tilesets/" + data_locations, TextFile.WriteOnly); // writes .oldata in directory of tileset image
+            locdataFile.write(locdata);
+            locdataFile.commit();
+        }
+
+        // terrain .oldata parsing
+        var landdata = "";
+        for (var i = 0; i < tilesets_dict["Landscape"].tiles.length; ++i) {
+            var tile      = tilesets_dict["Landscape"].tiles[i];
+            var tile_data = tile.properties();
+
+            if (Object.keys(tile_data).length > 0) { // skips tiles without properties
+                landdata += "[tile." + tile.id + "]\n";
+
+                for (var key in tile_data) {
+                    var val = typeof tile_data[key] === "string"  ? "\"" + tile_data[key] + "\""
+                            : typeof tile_data[key] === "number"  ? tile_data[key]
+                            : typeof tile_data[key] === "boolean" ? tile_data[key]
+                            : "";
+                    landdata += key + " = " + val + "\n";
+                }
+            }
+        }
+
+        // .oldata for terrain
+        if (landdata.length > 0) { // skips if no properties are set (= user wants to make .oldata manually)
+            var landdataFile = new TextFile(baseDir + "/../tilesets/" + data_terrain, TextFile.WriteOnly); // writes .oldata in directory of tileset image
+            landdataFile.write(landdata);
+            landdataFile.commit();
+        }
     },
 }